requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob
azure-storage-queue
azure-data-tables
azure-identity
python-telegram-bot==21.7
asyncio
aiohttp
openai==1.59.3
backoff==2.2.1


